+++ lib/constants.dart
@@ -33,6 +33,7 @@ const kMicromappingTapZoom = 19.0; // how much to zoom in when tapping a bunch o
 const kRotationThreshold = 30.0; // degrees, for snapping to zero rotation
 const kManualOption = '✍️'; // Emoji icon for entering values by hand
 const kMinElementsForWarning = 20000; // Alerting user when they have that many elements downloaded
+const kChangesetSplitGap = 0.02; // Decimal degrees, min distance between groups of changes
 
 // Should be exactly 8 lines in both lists.
 const kDefaultPresets = [