+++ CHANGELOG.md
@@ -24,6 +24,7 @@ _Unreleased_
 * Moved ATM, vending machines and parcel lockers to micromapping.
 * NSI suggestions for micromapping objects.
 * A warning when you've downloaded too much elements.
+* Changes are split in groups for uploading if possible.
 
 ## 0.6